bug fixes, refactoring

+++ createdb.sql
@@ -7,7 +7,7 @@ create table categories(
 create table expenses(
     id integer primary key,
     category_id integer,
-    ammount float,
+    ammount integer,
     created datetime,
     FOREIGN KEY(category_id) REFERENCES category(id)
 );
